server/requirements_fastapi.txt
# Multi-Model AI Search Server Requirements

# Core FastAPI dependencies
fastapi==0.115.12
uvicorn==0.34.3
pydantic>=2.0.0
python-multipart>=0.0.6  # For file uploads

# AI/ML dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers==4.52.4
open_clip_torch>=2.24.0
timm>=0.9.0

# Image processing
Pillow>=10.0.0

# Data handling
numpy>=1.24.0

# Database (Supabase + PostgreSQL)
supabase>=2.3.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0

# Development
python-dotenv>=1.0.0
